Moved hasJourneyClassification to CallbackValidator
