+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CallbackValidator.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd;
 
+import io.vavr.control.Try;
 import io.vavr.control.Validation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,7 +19,12 @@ import javax.annotation.Nonnull;
 import static io.vavr.control.Validation.invalid;
 import static io.vavr.control.Validation.valid;
 import static java.lang.String.format;
+import static java.util.Arrays.asList;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.YesNoFieldValues.YES;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.EXCEPTION;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.NEW_APPLICATION;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.SUPPLEMENTARY_EVIDENCE;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.SUPPLEMENTARY_EVIDENCE_WITH_OCR;
 
 @Component
 public class CallbackValidator {
@@ -28,6 +34,9 @@ public class CallbackValidator {
     private static final String CLASSIFICATION_SUPPLEMENTARY_EVIDENCE = "SUPPLEMENTARY_EVIDENCE";
     private static final String CLASSIFICATION_SUPPLEMENTARY_EVIDENCE_WITH_OCR = "SUPPLEMENTARY_EVIDENCE_WITH_OCR";
     private static final String CLASSIFICATION_EXCEPTION = "EXCEPTION";
+
+    private static final List<Classification> VALID_CLASSIFICATIONS_FOR_ATTACH_TO_CASE =
+            asList(EXCEPTION, SUPPLEMENTARY_EVIDENCE, SUPPLEMENTARY_EVIDENCE_WITH_OCR);
 
     private final CaseReferenceValidator caseReferenceValidator;
     private final ScannedDocumentValidator scannedDocumentValidator;
@@ -192,10 +201,41 @@ public class CallbackValidator {
                 .orElse(invalid("Missing poBox"));
     }
 
-    private boolean hasOcr(CaseDetails theCase) {
-        return getOcrData(theCase)
-                .map(list -> !CollectionUtils.isEmpty(list))
-                .orElse(false);
+    /**
+     * Used in createCase callback only.
+     * @param theCase from CCD
+     * @return Validation of it
+     */
+    public Validation<String, Classification> hasJourneyClassification(CaseDetails theCase) {
+        Optional<String> classificationOption = getJourneyClassification(theCase);
+
+        return classificationOption
+                .map(classification -> Try.of(() -> Classification.valueOf(classification)))
+                .map(Try::toValidation)
+                .map(validation -> validation
+                        .mapError(throwable -> "Invalid journeyClassification. Error: " + throwable.getMessage())
+                        .flatMap(classification -> validateClassification(classification, theCase))
+                )
+                .orElse(invalid("Missing journeyClassification"));
+    }
+
+    public Validation<String, Classification> hasJourneyClassificationForAttachToCase(CaseDetails theCase) {
+        Optional<String> classificationOption = getJourneyClassification(theCase);
+
+        return classificationOption
+                .map(classification -> Try.of(() -> Classification.valueOf(classification)))
+                .map(Try::toValidation)
+                .map(validation -> validation
+                        .mapError(throwable ->
+                                format(
+                                        "Journey Classification %s is not allowed "
+                                                + "when attaching exception record to a case",
+                                        classificationOption.get()
+                                )
+                        )
+                        .flatMap(classification -> validateClassificationForAttachToCase(classification, theCase))
+                )
+                .orElse(invalid("Missing journeyClassification"));
     }
 
     @SuppressWarnings("unchecked")
@@ -225,10 +265,58 @@ public class CallbackValidator {
                 .map(c -> (String) c);
     }
 
-    private static Optional<String> getAwaitingPaymentDcnProcessing(CaseDetails theCase) {
+    private Optional<String> getAwaitingPaymentDcnProcessing(CaseDetails theCase) {
         return Optional.ofNullable(theCase)
                 .map(CaseDetails::getData)
                 .map(data -> data.get("awaitingPaymentDCNProcessing"))
                 .map(c -> (String) c);
+    }
+
+    private Validation<String, Classification> validateClassification(
+            Classification classification,
+            CaseDetails theCase
+    ) {
+        if (SUPPLEMENTARY_EVIDENCE.equals(classification)) {
+            return invalid(format(
+                    "Event createNewCase not allowed for the current journey classification %s",
+                    classification
+            ));
+        }
+
+        if ((EXCEPTION.equals(classification) || NEW_APPLICATION.equals(classification)) && !hasOcr(theCase)) {
+            return invalid(format(
+                    "Event createNewCase not allowed for the current journey classification %s without OCR",
+                    classification
+            ));
+        }
+
+        return valid(classification);
+    }
+
+    private boolean hasOcr(CaseDetails theCase) {
+        return getOcrData(theCase)
+                .map(list -> !CollectionUtils.isEmpty(list))
+                .orElse(false);
+    }
+
+    private Validation<String, Classification> validateClassificationForAttachToCase(
+            Classification classification,
+            CaseDetails theCase
+    ) {
+        if (!VALID_CLASSIFICATIONS_FOR_ATTACH_TO_CASE.contains(classification)) {
+            return invalid(format(
+                    "The current Journey Classification %s is not allowed for attaching to case",
+                    classification
+            ));
+        }
+
+        if (SUPPLEMENTARY_EVIDENCE_WITH_OCR.equals(classification) && !hasOcr(theCase)) {
+            return invalid(format(
+                    "The current journey classification %s is not allowed without OCR data",
+                    classification
+            ));
+        }
+
+        return valid(classification);
     }
 }
