+++ src/main/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/callback/ExceptionRecordValidator.java
@@ -27,7 +27,6 @@ import static java.util.stream.Collectors.toList;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.FORMATTER;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.getOcrData;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.hasDateField;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidations.hasJourneyClassification;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.ENVELOPE_ID;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.EXCEPTION;
 
@@ -74,7 +73,8 @@ public class ExceptionRecordValidator {
         Validation<String, String> caseTypeIdValidation = callbackValidator.hasCaseTypeId(caseDetails);
         Validation<String, String> poBoxValidation = callbackValidator.hasPoBox(caseDetails);
         Validation<String, String> jurisdictionValidation = callbackValidator.hasJurisdiction(caseDetails);
-        Validation<String, Classification> journeyClassificationValidation = hasJourneyClassification(caseDetails);
+        Validation<String, Classification> journeyClassificationValidation =
+                callbackValidator.hasJourneyClassification(caseDetails);
 
         Validation<String, String> formTypeValidation = callbackValidator.hasFormType(caseDetails);
         // Exception journey classification may not have form type so skipping validation for it
