+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CallbackValidatorTest.java
@@ -31,6 +31,7 @@ import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.TestCaseBui
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.TestCaseBuilder.createCaseWith;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.OCR_DATA;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.EXCEPTION;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.NEW_APPLICATION;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.SUPPLEMENTARY_EVIDENCE;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.SUPPLEMENTARY_EVIDENCE_WITH_OCR;
 
@@ -419,6 +420,35 @@ class CallbackValidatorTest {
         // then
         assertThat(res.isInvalid()).isTrue();
         assertThat(res.getError()).isEqualTo("Missing poBox");
+    }
+
+    private static Object[][] classificationForAttachToCaseTestParams() {
+        return new Object[][]{
+                {"Invalid journey classification", createCaseWith(b -> b.data(ImmutableMap.of(JOURNEY_CLASSIFICATION, "invalid_classification"))), false, null, "Journey Classification invalid_classification is not allowed when attaching exception record to a case"},
+                {"Invalid journey classification", createCaseWith(b -> b.data(ImmutableMap.of(JOURNEY_CLASSIFICATION, NEW_APPLICATION.name()))), false, null, "The current Journey Classification NEW_APPLICATION is not allowed for attaching to case"},
+                {"Valid journey classification(supplementary evidence)", createCaseWith(b -> b.data(ImmutableMap.of(JOURNEY_CLASSIFICATION, SUPPLEMENTARY_EVIDENCE.name()))), true, SUPPLEMENTARY_EVIDENCE, null},
+                {"Valid journey classification(supplementary evidence with ocr)", createCaseWith(b -> b.data(ImmutableMap.of(JOURNEY_CLASSIFICATION, SUPPLEMENTARY_EVIDENCE_WITH_OCR.name(), OCR_DATA, asList(ImmutableMap.of("f1", "v1"))))), true, SUPPLEMENTARY_EVIDENCE_WITH_OCR, null},
+                {"Valid journey classification(exception)", createCaseWith(b -> b.data(ImmutableMap.of(JOURNEY_CLASSIFICATION, CLASSIFICATION_EXCEPTION))), true, EXCEPTION, null}
+        };
+    }
+
+    @ParameterizedTest(name = "{0}: valid:{2} value:{3} error:{4}")
+    @MethodSource("classificationForAttachToCaseTestParams")
+    @DisplayName("Should accept valid journey classification")
+    void canJourneyClassificationBeAttachedToCaseTest(
+            String caseDescription,
+            CaseDetails inputCase,
+            boolean valid,
+            Classification expectedValue,
+            String error
+    ) {
+        checkValidation(
+                inputCase,
+                valid,
+                expectedValue,
+                callbackValidator::hasJourneyClassificationForAttachToCase,
+                error
+        );
     }
 
     private static ImmutableMap<String, Object> caseDataWithOcr() {
