+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/callback/ExceptionRecordValidatorTest.java
@@ -40,6 +40,7 @@ import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.Te
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.JOURNEY_CLASSIFICATION;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.EXCEPTION;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.NEW_APPLICATION;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.servicebus.domains.envelopes.model.Classification.SUPPLEMENTARY_EVIDENCE_WITH_OCR;
 
 @ExtendWith(MockitoExtension.class)
 @SuppressWarnings("checkstyle:LineLength")
@@ -66,6 +67,8 @@ class ExceptionRecordValidatorTest {
         given(callbackValidator.hasJurisdiction(any())).willReturn(Validation.valid("BULKSCAN"));
         given(callbackValidator.hasAnId(any())).willReturn(Validation.valid(1234L));
         given(callbackValidator.hasPoBox(any())).willReturn(Validation.valid(PO_BOX));
+        given(callbackValidator.hasJourneyClassification(any(CaseDetails.class)))
+                .willReturn(Validation.valid(NEW_APPLICATION));
 
         // when
         var validation = exceptionRecordValidator.getValidation(validExceptionRecord);
@@ -86,6 +89,8 @@ class ExceptionRecordValidatorTest {
         given(callbackValidator.hasJurisdiction(any())).willReturn(Validation.valid(null));
         given(callbackValidator.hasAnId(any())).willReturn(Validation.valid(1234L));
         given(callbackValidator.hasPoBox(any())).willReturn(Validation.valid(PO_BOX));
+        given(callbackValidator.hasJourneyClassification(any(CaseDetails.class)))
+                .willReturn(Validation.valid(SUPPLEMENTARY_EVIDENCE_WITH_OCR));
 
         // when
         var validation = exceptionRecordValidator.getValidation(caseDetails);
@@ -139,6 +144,8 @@ class ExceptionRecordValidatorTest {
         given(callbackValidator.hasJurisdiction(any())).willReturn(Validation.valid(null));
         given(callbackValidator.hasAnId(any())).willReturn(Validation.valid(123L));
         given(callbackValidator.hasPoBox(any())).willReturn(Validation.valid(PO_BOX));
+        given(callbackValidator.hasJourneyClassification(any(CaseDetails.class)))
+                .willReturn(Validation.valid(SUPPLEMENTARY_EVIDENCE_WITH_OCR));
 
         var invalidOcrData = ImmutableList.of(
             ImmutableMap.of("value", ImmutableMap.of(
@@ -179,6 +186,8 @@ class ExceptionRecordValidatorTest {
         given(callbackValidator.hasJurisdiction(any())).willReturn(Validation.valid(null));
         given(callbackValidator.hasAnId(any())).willReturn(Validation.valid(1234L));
         given(callbackValidator.hasPoBox(any())).willReturn(Validation.valid(PO_BOX));
+        given(callbackValidator.hasJourneyClassification(any(CaseDetails.class)))
+                .willReturn(Validation.valid(SUPPLEMENTARY_EVIDENCE_WITH_OCR));
 
         // when
         var validation = exceptionRecordValidator.getValidation(input);
@@ -214,6 +223,8 @@ class ExceptionRecordValidatorTest {
         given(callbackValidator.hasJurisdiction(any())).willReturn(Validation.valid(null));
         given(callbackValidator.hasAnId(any())).willReturn(Validation.valid(1234L));
         given(callbackValidator.hasPoBox(any())).willReturn(Validation.valid(PO_BOX));
+        given(callbackValidator.hasJourneyClassification(any(CaseDetails.class)))
+                .willReturn(Validation.valid(SUPPLEMENTARY_EVIDENCE_WITH_OCR));
 
         checkValidationErrorMatches(
             input,
@@ -231,6 +242,8 @@ class ExceptionRecordValidatorTest {
         given(callbackValidator.hasJurisdiction(any())).willReturn(Validation.valid("BULKSCAN"));
         given(callbackValidator.hasAnId(any())).willReturn(Validation.valid(1234L));
         given(callbackValidator.hasPoBox(any())).willReturn(Validation.valid(PO_BOX));
+        given(callbackValidator.hasJourneyClassification(any(CaseDetails.class)))
+                .willReturn(Validation.valid(SUPPLEMENTARY_EVIDENCE_WITH_OCR));
 
         // then
         assertThatCode(() -> exceptionRecordValidator.getValidation(caseDetails))
